Added separate directory for affinity_group model since the original design of just throwing the affinity_group creation in the rider model did not work properly.
